+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # dev
 
+* Replace platform parameter with platforms in check(). (#497)
+
 * Update shortcut for Mac OS platform (@echasnovski, #393)
 
 * `cran_summary()` now works even if there is a NOTE/WARNING/ERROR in one platform and nothing on other platforms (@fabian-s, #259).